# src/UnitConversion/ModelShared/BundleFromOpenApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ModelShared (Service)",
    "description": "Data model containing model dependencies of the clients of the current microservice (WebApp, ServiceTest)",
    "version": "3.2.1"
  },
  "paths": {
...
    "/UnitSystem": {
      "get": {
        "tags": [
          "UnitSystem"
        ],
        "operationId": "GetAllUnitSystemId",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "UnitSystem"
        ],
        "operationId": "PostUnitSystem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OSDC.UnitConversion.Conversion.UnitSystem.DrillingEngineering.UnitSystem"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OSDC.UnitConversion.Conversion.UnitSystem.DrillingEngineering.UnitSystem"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OSDC.UnitConversion.Conversion.UnitSystem.DrillingEngineering.UnitSystem"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/UnitSystem/{id}": {
      "get": {
        "tags": [
          "UnitSystem"
        ],
        "operationId": "GetUnitSystemById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OSDC.UnitConversion.Conversion.UnitSystem.DrillingEngineering.UnitSystem"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "UnitSystem"
        ],
        "operationId": "PutUnitSystemById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OSDC.UnitConversion.Conversion.UnitSystem.DrillingEngineering.UnitSystem"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OSDC.UnitConversion.Conversion.UnitSystem.DrillingEngineering.UnitSystem"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OSDC.UnitConversion.Conversion.UnitSystem.DrillingEngineering.UnitSystem"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "UnitSystem"
        ],
        "operationId": "DeleteUnitSystemById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OSDC.UnitConversion.Conversion.UnitSystem.DrillingEngineering.UnitSystem": {
        "type": "object",
        "properties": {
          "Choices": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "ID": {
            "type": "string",
            "format": "uuid"
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Description": {
            "type": "string",
            "nullable": true
          },
          "IsDefault": {
            "type": "boolean"
          },
          "IsSI": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}